Change to use AssemblyInfo for getting version
