+++ Lab/2018/BuildSample/ZipReleaseConsole/Program.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
+using System.IO.Compression;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Reflection;
 
 namespace ZipReleaseConsole
 {
@@ -12,6 +13,21 @@ namespace ZipReleaseConsole
         {
             ZipHelper.CreateZipFileForAssembly(@"..\..\");
             ZipHelper.CreateZipFileForAssembly(@"..\..\..\UnitTest");
+        }
+
+        internal static string GetAssemblyFileVersion(string assemblyFilePath)
+        {
+            var assembly = Assembly.LoadFrom(assemblyFilePath);
+            var assemblyFileVersion = assembly.GetCustomAttribute<AssemblyFileVersionAttribute>();
+            return assemblyFileVersion != null ? assemblyFileVersion.Version : null;
+        }
+
+        public static void CreateZipFile(string inputDirPath, string outputZipFilePath)
+        {
+            var outputDirPath = Path.GetDirectoryName(outputZipFilePath);
+            Directory.CreateDirectory(outputDirPath);
+            File.Delete(outputZipFilePath);
+            ZipFile.CreateFromDirectory(inputDirPath, outputZipFilePath);
         }
     }
 }
